+++ tool/src/java/org/sakaiproject/evaluation/tool/utils/ItemBlockUtils.java
@@ -1,5 +1,5 @@
 /******************************************************************************
- * ItemBlockUtils.java - created by fengr@vt.edu on Jan 18, 2007
+ * ItemBlockUtils.java - created on Jan 18, 2007
  * 
  * Copyright (c) 2007 Virginia Polytechnic Institute and State University
  * Licensed under the Educational Community License version 1.0
@@ -9,25 +9,34 @@
  * 
  * Contributors:
  * Rui Feng (fengr@vt.edu)
+ * Aaron Zeckoski (aaronz@vt.edu) - project lead
  *****************************************************************************/
 
 package org.sakaiproject.evaluation.tool.utils;
 
 import java.util.ArrayList;
 import java.util.List;
-
 import org.sakaiproject.evaluation.model.EvalTemplateItem;
-/**
- *  includes static methods to handle BLOCK Child ietms
 
+// TODO - commenting and format corrected, still need to add error checking and ordering (items should be returned in sorted order)
+
+/**
+ * Includes static methods to handle BLOCK Child items
+ *  
  * @author Rui Feng (fengr@vt.edu)
+ * @author Aaron Zeckoski (aaronz@vt.edu)
  */
 public class ItemBlockUtils {
-	
-//	to filter out the Block child items, and only return non-child items
-	public static List getNonChildItems(List tempItemsList){
-		
+
+	/**
+	 * filter out the Block child items, and only return non-child items
+	 * 
+	 * @param tempItemsList a List of {@link EvalTemplateItem} objects in a template
+	 * @return a List of {@link EvalTemplateItem} objects without any block child objects
+	 */
+	public static List getNonChildItems(List tempItemsList){		
 		List nonChildItemsList = new ArrayList();
+
 		for(int i= 0; i< tempItemsList.size(); i++){
 			EvalTemplateItem tempItem1 = (EvalTemplateItem)tempItemsList.get(i);		
 			if(tempItem1.getBlockId()== null)
@@ -37,8 +46,14 @@ public class ItemBlockUtils {
 		return nonChildItemsList;
 	}
 
-//	return the child items which assocaited with the BlockParentId
-	public static List getChildItems(List tempItemsList, Integer blockParentId){
+	/**
+	 * return the child items which are associated with a block parent Id
+	 * 
+	 * @param tempItemsList a List of {@link EvalTemplateItem} objects in a template
+	 * @param blockParentId a unique identifier for an {@link EvalTemplateItem} which is a block parent
+	 * @return a List of {@link EvalTemplateItem} objects
+	 */
+	public static List getChildItems(List tempItemsList, Long blockParentId){
 		List childItemsList = new ArrayList();
 		
 		for(int i= 0; i< tempItemsList.size(); i++){
@@ -49,4 +64,5 @@ public class ItemBlockUtils {
 		
 		return childItemsList;
 	}
+
 }
